\name{pgMulticore-package}
\alias{pgMulticore-package}
\alias{pgMulticore}
\docType{package}
\title{
Low overhead multicore processing for BionavigatoR
}
\description{
Low overhead multicore processing designed for use with the Bionavigator-R interface. Will detect the number of cores on the local PC and separate e.g. a "per cell" process accordingly. Firewall permnission may be required.
}
\details{
\tabular{ll}{
Package: \tab pgMulticore\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-06-11\cr
License: \tab (c) PamGene 2013\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Who wrote it

Maintainer: Who to complain to <rdwijn@pamgene.com>

}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
